feat: streamline product routes and enhance SEO guidance

+++ routes/web.php
@@ -22,12 +22,13 @@ Route::get('/', function () {
 Route::prefix('admin')/* ->middleware(['auth']) */->group(function () {
     // products
     Route::get('products', [ProductController::class, 'index'])->name('admin.products.index');
-    Route::get('products-create', [ProductController::class, 'create'])->name('admin.products.create');
+    Route::get('products/create', [ProductController::class, 'create'])->name('admin.products.create');
     Route::post('products', [ProductController::class, 'store'])->name('admin.products.store');
-    Route::get('products/{id}', [ProductController::class, 'show']);
-    Route::put('products/{id}', [ProductController::class, 'update']);
+    Route::get('products/{id}', [ProductController::class, 'show'])->name('admin.products.show');
+    Route::get('products/{id}/edit', [ProductController::class, 'edit'])->name('admin.products.edit');
+    Route::put('products/{id}', [ProductController::class, 'update'])->name('admin.products.update');
     Route::patch('products/{id}', [ProductController::class, 'update']);
-    Route::delete('products/{id}', [ProductController::class, 'destroy']);
+    Route::delete('products/{id}', [ProductController::class, 'destroy'])->name('admin.products.destroy');
     Route::post('products/bulk', [ProductController::class, 'bulk'])->name('admin.products.bulk');
